Add files via upload
Task 1 
Inspect the test file provided, which is called shark-data.txt. This is a delimited text file which contains data about different types of sharks. Each new line represents a single shark species with seven fields delimited by the colon character.
Java program (as a NetBeans project) that opens up the data file, parses the shark data prints the following information to the console:
a) The three largest sharks by length, ordered by the largest first. Print the common name of each shark on a newline along with the maximum length in centimetres.
b) The three smallest sharks by length, ordered by the smallest first. Print the common name of each shark on a newline along with the maximum length in centimetres.
c) The total number of letters (not including whitespace characters) of all the Latin names for all sharks in the data file. Duplicate letters must be counted.
d) The total number of unique words contained within all of the Latin names for all sharks in the data file that have an even number of letters.
e) The total number of unique words contained within all of the Latin names for all sharks in the data file that have an odd number of letters.
On compiling and running, your program should have output similar to the following:
      --- Three largest sharks ---
      Shark A, Length = a centimetres
      Shark B, Length = b centimetres
      Shark C, Length = c centimetres
Where the values ‘Shark A’ through ‘Shark C’ and ‘Shark X’ through ‘Shark Z’ are replaced with the relevant shark common names, and the values a through c and x though z are replaced with their respective maximum lengths, and the values P, Q and R replaced with the respective letter and word counts.

Task 2
For this task you are required to write a Java program that analyses the oceanic regions associated with each shark listed in the test data file. For each unique region that is contained in the entire shark data set, the Java program should print to the output window (console) the oceanic region and an associated list of sharks (using their common name, in alphabetical order) that are known to roam within that region. Each oceanic region and associated list of sharks should be printed on a separate line, as illustrated below.
Region 1 -> [Shark A, Shark B, Shark C]
Region 2 -> [Shark D, Shark E]
Region 3 -> [Shark F, Shark G, Shark H, Shark I]
Each region should be listed only once, and the associated list of sharks for that region should not contain duplicates. Sharks should be listed using their common name. The printout should list regions alphabetically, and the list of sharks associated with each region should also be in alphabetical order (based on their common name).

Task 3
For this task, your Java program should prompt the User to enter a search string at the command prompt, i.e., the NetBeans console window. The program should then match the input search string against all of the Latin names in the shark data set and print a list of the sharks that match. The matching should be case insensitive. In other words, the strings “aBc” and “AbC” should match.
On entering a search string expression at the console (command prompt), and after processing the data matching, the program should return a list of matching sharks ordered by their common name in alphabetical order, but which also lists their Latin name with all occurrences of matching string in uppercase and the rest of the Latin name in lowercase. The following examples illustrate how the program should operate.
Example 1
Enter search string for Latin names: > tUm
You entered string "tUm"
The following matches have been found:-
Nurse Shark - Latin name: Ginglymostoma cirraTUM Variegated Shark - Latin name: Stegostoma fasciaTUM